Combine highlight and search into one module
